Add the dataplugin template subdirectory to the MANIFEST ... (#316)
* Update the package_data to include all template files

The command line tooling expects to find a manifest template at that
location in the site-packages

Due to the templates directory containing non-python files they won't be
automatically installed at the site-package upon pip install. Our current setup
wasn't looking for all template based files, which causes an exception when 
attempting to run the following command:
biothings-cli dataplugin create

Rather than using the MANIFEST file to control file distribution,
we control it through the setup.py because we plan to migrate 
from setup.py to pyproject.toml in the future and wish to eliminate
the MANIFEST file altogether. So we update the wildcard expression
to include a wider range of template based files
---------

Co-authored-by: Johnathan Schaff <jschaff@scripps.edu>