Update the elasticsearch query builder regex pattern  (#314)
* Revert "Tutorial update"

* Correct the construction argument shadowing ...

Update regex test and and test for verifying that the default
builder parser property matches one generated separately from
a builder instance

* Correct the construction argument shadowing ...

Update regex test and and test for verifying that the default
builder parser property matches one generated separately from
a builder instance

* Add the CURIE ID regex parsing test in isolation

* 0.12.x (#320)

* removed extra bullet point

* Remove line for switching branches now that docker-compose is in the main branch

* changed to main branch docker-compose url

* Revert "Merge branch 'master' into 0.12.x"

This reverts commit 209b8ab3cf207e3faf47de3f848b15e0bff4ff0b, reversing
changes made to 738af0a965ed8108602f2a532746170a14d1e487.

---------

Co-authored-by: Jason Lin <35415519+jal347@users.noreply.github.com>

* Include non-word character matching in scope regex

* Add the default regex pattern verification method

* Correct the pattern typo in the regex verification

* Add regex ordering tests

* Flush out the regex ordering tests

* Remove the moved test

* Add updated comments / docstring documentation

* Move the default regex pattern to settings.default

* Move the regex pattern builder to QStringParser ...

Makes more sense to move it to be a private method of the parser as the
Builder class doesn't even store the patterns. It keeps an instance of
the parser after construction so it shouldn't have to care about how the
patterns are built.

* Adopt the regex tests to the change ...

Due to the change in default regex, we had to change the scoping of the
expected query for the "a:b:c" query as we now denote that that the
scope is "a" with the term being "b:c" with this new default regex

* Replace module access from default to config

* Replace module access from default to config

* Add the default regex pattern to the ESQueryBuilder constructor ...

Took some exploring because of the initial confusion between the web
configuration and the hub configuration as these both can appear quite
similar and use different mechnanisms for loading. Note that the hub
configuration uses the utils.configuration.ConfigurationWrapper
object versus the web configuration uses the
web.settings.configs.ConfigModule

* Typo correction

* Change the default pattern to the configuration value

* Update the tests to adjust for changes

* Correct the codacy linting messages

* Remove implicit concat produced by black

---------

Co-authored-by: Jason Lin <35415519+jal347@users.noreply.github.com>
Co-authored-by: Johnathan Schaff <jschaff@scripps.edu>
Co-authored-by: Chunlei Wu <anewgene@yahoo.com>
Co-authored-by: Everaldo <everaldorodrigo@gmail.com>