Add id checking to IgnoreDuplicatedStorage ...
For the sequential case, this storage works as expected because
each document is checked for uniqueness. When processing this in batch,
there needs to be some more filtering involved especially if the number
of document is small with a higher ratio of duplicates in those
documents.

If you attempt to upload to the database and each batch has duplicates
in it, the entire batch will continually get thrown out until no
documents were actually uploaded to the database. Instead if we verify
the uniqueness constraint prior to uploading by ensuring we have a
one-to-one ratio of id to documents in our collection, then we can
ensure a safe upload.

This still discards identifical or near identical documents without
throwing out an entire batch of potentially valid documents in the batch
upload