Move the manager instances into the hub module ...
The grouping of the manager instances in utils with `BaseManager`,
`BaseSourceManager` and `JobManager` makes it seem like all three are
releated. However, the `BaseManager` and `BaseSourceManager` are only
used as parent classes for instantiating different instances within
the dataload and dataplugin hub packages. It makes more sense to
colocate them together.