Separate the web server start commands ...
Separates the command for starting the hub web api
and the ssh hub. Allows for easier testing of the
specific hub server when making a function call.

Also utilizes self.features in place of the class property
DEFAULT_FEATURES when configuring the hub server so that
certain features aren't assumed to be required when they were
not passed. Useful when making a minimal hub server for testing
purposes