Plugin directory loading error (#388)
* Check for the datafolder existance

* Update warning messaging

* Separate the web server start commands ...

Separates the command for starting the hub web api
and the ssh hub. Allows for easier testing of the
specific hub server when making a function call.

Also utilizes self.features in place of the class property
DEFAULT_FEATURES when configuring the hub server so that
certain features aren't assumed to be required when they were
not passed. Useful when making a minimal hub server for testing
purposes

* Re-add the deleted pickle.dump and function call

* Move the configuration to the beginning of the start method

* Move the Assisted classes to dataload ...

Due to the limitations of pickle, we can't have these classes be
submodules of another module. Another reason to migrate from pickle

* Update the Assisted class access in dataload

* Move template generation to the root of dataplugin module

* Add the metaclass to globals stack

* Black formatting

---------

Co-authored-by: jschaff <jschaff@scripps.edu>