Data source registration cleanup (#361)
* Add the data_source path to the managers in cli

* Move the manager instances into the hub module ...

The grouping of the manager instances in utils with `BaseManager`,
`BaseSourceManager` and `JobManager` makes it seem like all three are
releated. However, the `BaseManager` and `BaseSourceManager` are only
used as parent classes for instantiating different instances within
the dataload and dataplugin hub packages. It makes more sense to
colocate them together.

* Use the new paths for the manager parent classes

* Cleanup some of the assistant implementation

* Codacy warnings cleanup

---------

Co-authored-by: jschaff <jschaff@scripps.edu>