Overhaul the configuration_wrapper testing ...
Update the conftest to hold the fixtures releated to
testing the configuration_wrapper, which is what this test module
is actually doing. Somewhat differently worded from it's original
which makes it feel like it was just config testing. Combines
all the tests into one file, and creates a data folder for the two
different configurations to load via fixture