Complete overhaul for the test configuration loading ...
Switch to a sessionstart hook so that the configuration is loaded
before we attempt to do anything for the tests. This will load the newly
created TestConfig which a datatype defined specifically for testing.
It provides a way for us to load and override configuration parameters
depending on what test module we're located in. This also moves all of
the logic from the `test_config.py` at the root of the tests directory,
so we have one less file to manage