Modifications to the dynamic uploader template ...
Changes the import structure for the uploaders. Explicitly sets the
dataplugin directory as the sys.path entry to avoid using the current
directory and potentially leverage the wrong directory. Also attempts to
reload the parser function before importing the function to ensure we
load the correct function. Cleans up the function for extracting the
code from a manifest defined module function