Remove manifest-file option from validate command ...
Removes the option to specify a manifest file directly.
This simplifies our logic for handling the plugin pathing.
Also provides a check to see if the plugin-name is provided
in SINGULAR mode, in which case it is ignored to ensure that
we don't modify the plugin path erroneously. The plugin-name
is primarily used for HUB mode