Various bugfixes and cleanup (#404)
* Change validate argument from --plugin to --plugin-name

* Fix the List typing for actual list invocations

* Remove manifest-file option from validate command ...

Removes the option to specify a manifest file directly.
This simplifies our logic for handling the plugin pathing.
Also provides a check to see if the plugin-name is provided
in SINGULAR mode, in which case it is ignored to ensure that
we don't modify the plugin path erroneously. The plugin-name
is primarily used for HUB mode

* Add exception handling for dumper set_release call

* Clean-up the last modified header warning message