Isolate cli job management (#405)
* Revert to using CLIJobManager ...

Due to the "spawn" versus "fork" paradigm,
we want to ensure that the command-line has wide support
over full parity of the hub. So until we re-factor the
hub configuration to support changing the multiprocessing
paradigm, we'll leverage a simplified versus of the JobManager

* Remove duplicate -m shorthand arg for inspect command

* Add support for multiple uploaders to index command ...

Unexpected found this while testing the tutorials plugin, but
the index command never handle multiple uploaders appropriately.

We needed to add status registration calls to the upload process
so that the upload state would be created in the biothing_hubdb.src_dump
table / collection. After that the plugin_name.subsource_name would be
resolvable via the `get_source_fullname` function that attempts
to look under `upload.jobs` field within the biothings_hubdb.src_dump
entries

Added additional error handling to this and the inspect command to
ensure parity between the two with messaging, along with producing more
user friendly output upon forgetting to provide --sub-source-name

We also build a different job manager for the indexing task

* Add warning for non-MongoDB usage with index command